Rename algo_selection to algorithms.
