+++ pyproject.toml
@@ -151,7 +151,7 @@ extend-ignore = [
 "docs/source/conf.py" = ["E501", "ERA001", "DTZ005"]
 "src/optimagic/parameters/kernel_transformations.py" = ["ARG001", "N806"]
 "docs/source/*" = ["B018"]
-"src/optimagic/algo_selection.py" = ["E501"]
+"src/optimagic/algorithms.py" = ["E501"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "google"
