+++ src/optimagic/algorithms.py
