+++ src/optimagic/optimization/create_optimization_problem.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import Any, Callable, Type
 
 from optimagic import deprecations
-from optimagic.algo_selection import ALL_ALGORITHMS
+from optimagic.algorithms import ALL_ALGORITHMS
 from optimagic.deprecations import (
     handle_log_options_throw_deprecated_warning,
     replace_and_warn_about_deprecated_algo_options,
